+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   - Photographs can no longer be placed on the wall. Existing hanging photographs will not be removed.
   
 - Lightroom comparator output now based on the number of Photographs in the output slot rather than the selected frame.
+- Chromatic Process can now be enabled for all black and white images by placing an Amethyst Cluster on top of a Lightroom.
 
 ---